Revert "use big circle for loading"
This reverts commit 7855c287b0b076a4f28753fe73534244e231076e.
